Move host impls to their own module
