+++ crates/factor-variables/src/lib.rs
@@ -1,3 +1,4 @@
+mod host;
 pub mod runtime_config;
 pub mod spin_cli;
 
@@ -9,7 +10,6 @@ use spin_factors::{
     anyhow, ConfigureAppContext, Factor, InitContext, InstanceBuilders, PrepareContext,
     RuntimeFactors, SelfInstanceBuilder,
 };
-use spin_world::{async_trait, v1, v2::variables};
 
 /// A factor for providing variables to components.
 pub struct VariablesFactor {
@@ -87,45 +87,3 @@ impl InstanceState {
 }
 
 impl SelfInstanceBuilder for InstanceState {}
-
-#[async_trait]
-impl variables::Host for InstanceState {
-    async fn get(&mut self, key: String) -> Result<String, variables::Error> {
-        let key = spin_expressions::Key::new(&key).map_err(expressions_to_variables_err)?;
-        self.expression_resolver
-            .resolve(&self.component_id, key)
-            .await
-            .map_err(expressions_to_variables_err)
-    }
-
-    fn convert_error(&mut self, error: variables::Error) -> anyhow::Result<variables::Error> {
-        Ok(error)
-    }
-}
-
-#[async_trait]
-impl v1::config::Host for InstanceState {
-    async fn get_config(&mut self, key: String) -> Result<String, v1::config::Error> {
-        <Self as variables::Host>::get(self, key)
-            .await
-            .map_err(|err| match err {
-                variables::Error::InvalidName(msg) => v1::config::Error::InvalidKey(msg),
-                variables::Error::Undefined(msg) => v1::config::Error::Provider(msg),
-                other => v1::config::Error::Other(format!("{other}")),
-            })
-    }
-
-    fn convert_error(&mut self, err: v1::config::Error) -> anyhow::Result<v1::config::Error> {
-        Ok(err)
-    }
-}
-
-fn expressions_to_variables_err(err: spin_expressions::Error) -> variables::Error {
-    use spin_expressions::Error;
-    match err {
-        Error::InvalidName(msg) => variables::Error::InvalidName(msg),
-        Error::Undefined(msg) => variables::Error::Undefined(msg),
-        Error::Provider(err) => variables::Error::Provider(err.to_string()),
-        other => variables::Error::Other(format!("{other}")),
-    }
-}
